fix(Validate) : boolean -> void

+++ src/main/java/lotto/CreateBonusNumber.java
@@ -3,9 +3,9 @@ package lotto;
 import java.util.List;
 
 public class CreateBonusNumber {
-    public int getBonusNumber(List<Integer> winningNumber, String userInput){
+    public long getBonusNumber(List<Integer> winningNumber, String userInput){
         Validate checkBonusNumber = new Validate();
         checkBonusNumber.validateBonusNumber(winningNumber, userInput);
-        return Integer.parseInt(userInput);
+        return Long.parseLong(userInput);
     }
 }
