+++ src/main/java/lotto/GetRateOfReturn.java
@@ -1,8 +1,8 @@
 package lotto;
 
 public class GetRateOfReturn {
-    public double calculateRate(int [] result, int money){
-        int value = result[2] * 5000 + result[3] * 50000 + result[4] * 150000 + result[5] * 2000000000 + result[6] * 30000000;
+    public double calculateRate(long [] result, long money){
+        long value = result[2] * 5000 + result[3] * 50000 + result[4] * 150000 + result[5] * 2000000000 + result[6] * 30000000;
         double coast = value / (double)money;
         return ((double) Math.round(coast * 1000) / 10);
     }
