+++ src/main/java/lotto/PrintMessenger.java
@@ -6,7 +6,7 @@ public class PrintMessenger {
     public static void askPriceToBuy(){
         System.out.println("구입금액을 입력해 주세요.");
     }
-    public static void volumeBuying(int volume){
+    public static void volumeBuying(long volume){
         System.out.println(volume + "개를 구매했습니다.");
     }
     public static void askWinningNumber()
@@ -16,7 +16,7 @@ public class PrintMessenger {
     public static void askBonusNumber(){
         System.out.println("보너스 번호를 입력해 주세요.");
     }
-    public static void winningResult(int [] countResult)// list를 받을지 int를 받을지 고민;;
+    public static void winningResult(long [] countResult)// list를 받을지 int를 받을지 고민;;
     {
         System.out.println("당첨 통계\n" + "---");
         System.out.println("3개 일치 (5,000원) -" + countResult[3]+ "개");
